+++ macos/Classes/frb_generated.h
@@ -86,6 +86,14 @@ typedef struct wire_cst_tx_fee {
   union TxFeeKind kind;
 } wire_cst_tx_fee;
 
+typedef struct wire_cst_swap_master_key {
+  struct wire_cst_list_prim_u_8_strict *xprv;
+  struct wire_cst_list_prim_u_8_strict *xpub;
+  int32_t network;
+  struct wire_cst_list_prim_u_8_strict *mnemonic;
+  struct wire_cst_list_prim_u_8_strict *fingerprint;
+} wire_cst_swap_master_key;
+
 typedef struct wire_cst_l_btc_swap_script_str {
   int32_t swap_type;
   struct wire_cst_list_prim_u_8_strict *funding_addrs;
@@ -142,9 +150,62 @@ typedef struct wire_cst_lnurl {
   struct wire_cst_list_prim_u_8_strict *value;
 } wire_cst_lnurl;
 
+typedef struct wire_cst_transaction {
+  struct wire_cst_list_prim_u_8_strict *id;
+  struct wire_cst_list_prim_u_8_strict *hex;
+  uint32_t *eta;
+} wire_cst_transaction;
+
+typedef struct wire_cst_swap_status_response {
+  int32_t status;
+  struct wire_cst_transaction *transaction;
+  struct wire_cst_list_prim_u_8_strict *failure_reason;
+  struct wire_cst_list_prim_u_8_strict *error;
+} wire_cst_swap_status_response;
+
+typedef struct wire_cst_swap_stream_status {
+  struct wire_cst_list_prim_u_8_strict *id;
+  int32_t status;
+  struct wire_cst_list_prim_u_8_strict *error;
+  struct wire_cst_transaction *transaction;
+} wire_cst_swap_stream_status;
+
+typedef struct wire_cst_list_btc_ln_swap {
+  struct wire_cst_btc_ln_swap *ptr;
+  int32_t len;
+} wire_cst_list_btc_ln_swap;
+
+typedef struct wire_cst_list_chain_swap {
+  struct wire_cst_chain_swap *ptr;
+  int32_t len;
+} wire_cst_list_chain_swap;
+
+typedef struct wire_cst_list_lbtc_ln_swap {
+  struct wire_cst_lbtc_ln_swap *ptr;
+  int32_t len;
+} wire_cst_list_lbtc_ln_swap;
+
+typedef struct wire_cst_restored_swap_summary {
+  struct wire_cst_list_prim_u_8_strict *id;
+  int32_t kind;
+  struct wire_cst_list_prim_u_8_strict *status;
+  uint64_t created_at;
+  struct wire_cst_list_prim_u_8_strict *from;
+  struct wire_cst_list_prim_u_8_strict *to;
+  uint64_t amount;
+  bool recoverable;
+} wire_cst_restored_swap_summary;
+
+typedef struct wire_cst_list_restored_swap_summary {
+  struct wire_cst_restored_swap_summary *ptr;
+  int32_t len;
+} wire_cst_list_restored_swap_summary;
+
 typedef struct wire_cst_swap_limits {
   uint64_t minimal;
   uint64_t maximal;
+  uint64_t *maximal_zero_conf;
+  uint64_t *minimal_batched;
 } wire_cst_swap_limits;
 
 typedef struct wire_cst_chain_swap_fees {
@@ -155,10 +216,10 @@ typedef struct wire_cst_chain_swap_fees {
 } wire_cst_chain_swap_fees;
 
 typedef struct wire_cst_chain_fees_and_limits {
-  struct wire_cst_swap_limits btc_limits;
-  struct wire_cst_swap_limits lbtc_limits;
-  struct wire_cst_chain_swap_fees btc_fees;
-  struct wire_cst_chain_swap_fees lbtc_fees;
+  struct wire_cst_swap_limits lbtc_to_btc_limits;
+  struct wire_cst_swap_limits btc_to_lbtc_limits;
+  struct wire_cst_chain_swap_fees lbtc_to_btc_fees;
+  struct wire_cst_chain_swap_fees btc_to_lbtc_fees;
 } wire_cst_chain_fees_and_limits;
 
 typedef struct wire_cst_decoded_invoice {
@@ -265,8 +326,7 @@ void frbgen_boltz_wire__crate__api__btc_ln__btc_ln_swap_new(int64_t port_,
                                                             struct wire_cst_list_prim_u_8_strict *referral_id);
 
 void frbgen_boltz_wire__crate__api__btc_ln__btc_ln_swap_new_reverse(int64_t port_,
-                                                                    struct wire_cst_list_prim_u_8_strict *mnemonic,
-                                                                    struct wire_cst_list_prim_u_8_strict *passphrase,
+                                                                    struct wire_cst_swap_master_key *swap_master_key,
                                                                     uint64_t index,
                                                                     uint64_t out_amount,
                                                                     struct wire_cst_list_prim_u_8_strict *out_address,
@@ -277,8 +337,7 @@ void frbgen_boltz_wire__crate__api__btc_ln__btc_ln_swap_new_reverse(int64_t port
                                                                     struct wire_cst_list_prim_u_8_strict *referral_id);
 
 void frbgen_boltz_wire__crate__api__btc_ln__btc_ln_swap_new_submarine(int64_t port_,
-                                                                      struct wire_cst_list_prim_u_8_strict *mnemonic,
-                                                                      struct wire_cst_list_prim_u_8_strict *passphrase,
+                                                                      struct wire_cst_swap_master_key *swap_master_key,
                                                                       uint64_t index,
                                                                       struct wire_cst_list_prim_u_8_strict *invoice,
                                                                       int32_t network,
@@ -366,8 +425,7 @@ void frbgen_boltz_wire__crate__api__chain_swap__chain_swap_new(int64_t port_,
 
 void frbgen_boltz_wire__crate__api__chain_swap__chain_swap_new_swap(int64_t port_,
                                                                     int32_t direction,
-                                                                    struct wire_cst_list_prim_u_8_strict *mnemonic,
-                                                                    struct wire_cst_list_prim_u_8_strict *passphrase,
+                                                                    struct wire_cst_swap_master_key *swap_master_key,
                                                                     uint64_t index,
                                                                     uint64_t amount,
                                                                     bool is_testnet,
@@ -402,9 +460,9 @@ void frbgen_boltz_wire__crate__api__transactions__check_vout_0_outspend(int64_t 
                                                                         struct wire_cst_list_prim_u_8_strict *boltz_url,
                                                                         int32_t *chain_swap_direction);
 
-void frbgen_boltz_wire__crate__api__types__decoded_invoice_from_string(int64_t port_,
-                                                                       struct wire_cst_list_prim_u_8_strict *s,
-                                                                       struct wire_cst_list_prim_u_8_strict *boltz_url);
+void frbgen_boltz_wire__crate__api__invoice__decoded_invoice_from_string(int64_t port_,
+                                                                         struct wire_cst_list_prim_u_8_strict *s,
+                                                                         struct wire_cst_list_prim_u_8_strict *boltz_url);
 
 void frbgen_boltz_wire__crate__api__fees__fees_chain(int64_t port_, struct wire_cst_fees *that);
 
@@ -414,17 +472,6 @@ void frbgen_boltz_wire__crate__api__fees__fees_new(int64_t port_,
 void frbgen_boltz_wire__crate__api__fees__fees_reverse(int64_t port_, struct wire_cst_fees *that);
 
 void frbgen_boltz_wire__crate__api__fees__fees_submarine(int64_t port_, struct wire_cst_fees *that);
-
-void frbgen_boltz_wire__crate__api__types__key_pair_generate(int64_t port_,
-                                                             struct wire_cst_list_prim_u_8_strict *mnemonic,
-                                                             struct wire_cst_list_prim_u_8_strict *passphrase,
-                                                             int32_t network,
-                                                             uint64_t index,
-                                                             int32_t swap_type);
-
-void frbgen_boltz_wire__crate__api__types__key_pair_new(int64_t port_,
-                                                        struct wire_cst_list_prim_u_8_strict *secret_key,
-                                                        struct wire_cst_list_prim_u_8_strict *public_key);
 
 void frbgen_boltz_wire__crate__api__types__l_btc_swap_script_str_new(int64_t port_,
                                                                      int32_t swap_type,
@@ -486,8 +533,7 @@ void frbgen_boltz_wire__crate__api__lbtc_ln__lbtc_ln_swap_new(int64_t port_,
                                                               struct wire_cst_list_prim_u_8_strict *referral_id);
 
 void frbgen_boltz_wire__crate__api__lbtc_ln__lbtc_ln_swap_new_reverse(int64_t port_,
-                                                                      struct wire_cst_list_prim_u_8_strict *mnemonic,
-                                                                      struct wire_cst_list_prim_u_8_strict *passphrase,
+                                                                      struct wire_cst_swap_master_key *swap_master_key,
                                                                       uint64_t index,
                                                                       uint64_t out_amount,
                                                                       struct wire_cst_list_prim_u_8_strict *out_address,
@@ -498,8 +544,7 @@ void frbgen_boltz_wire__crate__api__lbtc_ln__lbtc_ln_swap_new_reverse(int64_t po
                                                                       struct wire_cst_list_prim_u_8_strict *referral_id);
 
 void frbgen_boltz_wire__crate__api__lbtc_ln__lbtc_ln_swap_new_submarine(int64_t port_,
-                                                                        struct wire_cst_list_prim_u_8_strict *mnemonic,
-                                                                        struct wire_cst_list_prim_u_8_strict *passphrase,
+                                                                        struct wire_cst_swap_master_key *swap_master_key,
                                                                         uint64_t index,
                                                                         struct wire_cst_list_prim_u_8_strict *invoice,
                                                                         int32_t network,
@@ -539,12 +584,60 @@ void frbgen_boltz_wire__crate__api__lnurl__lnurl_withdraw(int64_t port_,
                                                           struct wire_cst_lnurl *that,
                                                           struct wire_cst_list_prim_u_8_strict *invoice);
 
-void frbgen_boltz_wire__crate__api__types__pre_image_generate(int64_t port_);
+void frbgen_boltz_wire__crate__api__secrets__pre_image_from_invoice_str(int64_t port_,
+                                                                        struct wire_cst_list_prim_u_8_strict *invoice);
 
-void frbgen_boltz_wire__crate__api__types__pre_image_new(int64_t port_,
-                                                         struct wire_cst_list_prim_u_8_strict *value,
-                                                         struct wire_cst_list_prim_u_8_strict *sha256,
-                                                         struct wire_cst_list_prim_u_8_strict *hash160);
+void frbgen_boltz_wire__crate__api__secrets__pre_image_new(int64_t port_,
+                                                           struct wire_cst_list_prim_u_8_strict *value,
+                                                           struct wire_cst_list_prim_u_8_strict *sha256,
+                                                           struct wire_cst_list_prim_u_8_strict *hash160);
+
+void frbgen_boltz_wire__crate__api__restore__restore_chain_swaps(int64_t port_,
+                                                                 struct wire_cst_swap_master_key *swap_master_key,
+                                                                 struct wire_cst_list_prim_u_8_strict *btc_electrum_url,
+                                                                 struct wire_cst_list_prim_u_8_strict *lbtc_electrum_url,
+                                                                 struct wire_cst_list_prim_u_8_strict *boltz_url);
+
+void frbgen_boltz_wire__crate__api__restore__restore_ln_btc_swaps(int64_t port_,
+                                                                  struct wire_cst_swap_master_key *swap_master_key,
+                                                                  struct wire_cst_list_prim_u_8_strict *electrum_url,
+                                                                  struct wire_cst_list_prim_u_8_strict *boltz_url);
+
+void frbgen_boltz_wire__crate__api__restore__restore_ln_lbtc_swaps(int64_t port_,
+                                                                   struct wire_cst_swap_master_key *swap_master_key,
+                                                                   struct wire_cst_list_prim_u_8_strict *electrum_url,
+                                                                   struct wire_cst_list_prim_u_8_strict *boltz_url);
+
+void frbgen_boltz_wire__crate__api__restore__restore_swap_index(int64_t port_,
+                                                                struct wire_cst_swap_master_key *swap_master_key,
+                                                                struct wire_cst_list_prim_u_8_strict *boltz_url);
+
+void frbgen_boltz_wire__crate__api__restore__restore_swap_summaries(int64_t port_,
+                                                                    struct wire_cst_swap_master_key *swap_master_key,
+                                                                    struct wire_cst_list_prim_u_8_strict *boltz_url);
+
+void frbgen_boltz_wire__crate__api__secrets__swap_master_key_create(int64_t port_,
+                                                                    struct wire_cst_list_prim_u_8_strict *wallet_mnemonic,
+                                                                    struct wire_cst_list_prim_u_8_strict *wallet_passphrase,
+                                                                    int32_t network);
+
+WireSyncRust2DartDco frbgen_boltz_wire__crate__api__swap_status__swap_status_as_string(int32_t that);
+
+WireSyncRust2DartDco frbgen_boltz_wire__crate__api__swap_status__swap_status_from_json_string(struct wire_cst_list_prim_u_8_strict *status);
+
+WireSyncRust2DartDco frbgen_boltz_wire__crate__api__swap_status__swap_status_response_from_json(struct wire_cst_list_prim_u_8_strict *json);
+
+WireSyncRust2DartDco frbgen_boltz_wire__crate__api__swap_status__swap_status_response_to_json(struct wire_cst_swap_status_response *that);
+
+WireSyncRust2DartDco frbgen_boltz_wire__crate__api__swap_status__swap_status_to_json_string(int32_t that);
+
+WireSyncRust2DartDco frbgen_boltz_wire__crate__api__swap_status__swap_stream_status_from_json(struct wire_cst_list_prim_u_8_strict *json);
+
+WireSyncRust2DartDco frbgen_boltz_wire__crate__api__swap_status__swap_stream_status_to_json(struct wire_cst_swap_stream_status *that);
+
+WireSyncRust2DartDco frbgen_boltz_wire__crate__api__swap_status__transaction_from_json(struct wire_cst_list_prim_u_8_strict *json);
+
+WireSyncRust2DartDco frbgen_boltz_wire__crate__api__swap_status__transaction_to_json(struct wire_cst_transaction *that);
 
 struct wire_cst_boltz_error *frbgen_boltz_cst_new_box_autoadd_boltz_error(void);
 
@@ -572,11 +665,29 @@ struct wire_cst_pre_image *frbgen_boltz_cst_new_box_autoadd_pre_image(void);
 
 int32_t *frbgen_boltz_cst_new_box_autoadd_side(int32_t value);
 
+struct wire_cst_swap_master_key *frbgen_boltz_cst_new_box_autoadd_swap_master_key(void);
+
+struct wire_cst_swap_status_response *frbgen_boltz_cst_new_box_autoadd_swap_status_response(void);
+
+struct wire_cst_swap_stream_status *frbgen_boltz_cst_new_box_autoadd_swap_stream_status(void);
+
+struct wire_cst_transaction *frbgen_boltz_cst_new_box_autoadd_transaction(void);
+
 struct wire_cst_tx_fee *frbgen_boltz_cst_new_box_autoadd_tx_fee(void);
+
+uint32_t *frbgen_boltz_cst_new_box_autoadd_u_32(uint32_t value);
 
 uint64_t *frbgen_boltz_cst_new_box_autoadd_u_64(uint64_t value);
 
+struct wire_cst_list_btc_ln_swap *frbgen_boltz_cst_new_list_btc_ln_swap(int32_t len);
+
+struct wire_cst_list_chain_swap *frbgen_boltz_cst_new_list_chain_swap(int32_t len);
+
+struct wire_cst_list_lbtc_ln_swap *frbgen_boltz_cst_new_list_lbtc_ln_swap(int32_t len);
+
 struct wire_cst_list_prim_u_8_strict *frbgen_boltz_cst_new_list_prim_u_8_strict(int32_t len);
+
+struct wire_cst_list_restored_swap_summary *frbgen_boltz_cst_new_list_restored_swap_summary(int32_t len);
 static int64_t dummy_method_to_enforce_bundling(void) {
     int64_t dummy_var = 0;
     dummy_var ^= ((int64_t) (void*) frbgen_boltz_cst_new_box_autoadd_boltz_error);
@@ -592,9 +703,18 @@ static int64_t dummy_method_to_enforce_bundling(void) {
     dummy_var ^= ((int64_t) (void*) frbgen_boltz_cst_new_box_autoadd_lnurl);
     dummy_var ^= ((int64_t) (void*) frbgen_boltz_cst_new_box_autoadd_pre_image);
     dummy_var ^= ((int64_t) (void*) frbgen_boltz_cst_new_box_autoadd_side);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_cst_new_box_autoadd_swap_master_key);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_cst_new_box_autoadd_swap_status_response);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_cst_new_box_autoadd_swap_stream_status);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_cst_new_box_autoadd_transaction);
     dummy_var ^= ((int64_t) (void*) frbgen_boltz_cst_new_box_autoadd_tx_fee);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_cst_new_box_autoadd_u_32);
     dummy_var ^= ((int64_t) (void*) frbgen_boltz_cst_new_box_autoadd_u_64);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_cst_new_list_btc_ln_swap);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_cst_new_list_chain_swap);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_cst_new_list_lbtc_ln_swap);
     dummy_var ^= ((int64_t) (void*) frbgen_boltz_cst_new_list_prim_u_8_strict);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_cst_new_list_restored_swap_summary);
     dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__btc_ln__btc_ln_swap_broadcast_boltz);
     dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__btc_ln__btc_ln_swap_broadcast_local);
     dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__btc_ln__btc_ln_swap_claim);
@@ -627,6 +747,7 @@ static int64_t dummy_method_to_enforce_bundling(void) {
     dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__fees__fees_new);
     dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__fees__fees_reverse);
     dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__fees__fees_submarine);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__invoice__decoded_invoice_from_string);
     dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__lbtc_ln__lbtc_ln_swap_broadcast_boltz);
     dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__lbtc_ln__lbtc_ln_swap_broadcast_local);
     dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__lbtc_ln__lbtc_ln_swap_claim);
@@ -646,14 +767,26 @@ static int64_t dummy_method_to_enforce_bundling(void) {
     dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__lnurl__lnurl_new);
     dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__lnurl__lnurl_validate);
     dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__lnurl__lnurl_withdraw);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__restore__restore_chain_swaps);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__restore__restore_ln_btc_swaps);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__restore__restore_ln_lbtc_swaps);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__restore__restore_swap_index);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__restore__restore_swap_summaries);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__secrets__pre_image_from_invoice_str);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__secrets__pre_image_new);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__secrets__swap_master_key_create);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__swap_status__swap_status_as_string);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__swap_status__swap_status_from_json_string);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__swap_status__swap_status_response_from_json);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__swap_status__swap_status_response_to_json);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__swap_status__swap_status_to_json_string);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__swap_status__swap_stream_status_from_json);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__swap_status__swap_stream_status_to_json);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__swap_status__transaction_from_json);
+    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__swap_status__transaction_to_json);
     dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__transactions__check_vout_0_outspend);
     dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__types__btc_swap_script_str_new);
-    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__types__decoded_invoice_from_string);
-    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__types__key_pair_generate);
-    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__types__key_pair_new);
     dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__types__l_btc_swap_script_str_new);
-    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__types__pre_image_generate);
-    dummy_var ^= ((int64_t) (void*) frbgen_boltz_wire__crate__api__types__pre_image_new);
     dummy_var ^= ((int64_t) (void*) store_dart_post_cobject);
     return dummy_var;
 }